+++ aimmo-game/simulation/avatar/avatar_view.py
@@ -1,5 +1,5 @@
 from simulation.geography.location import Location
-from simulation.direction import NORTH, SOUTH, EAST, WEST
+from simulation.geography.direction import NORTH, SOUTH, EAST, WEST
 
 
 class AvatarView:
