+++ aimmo-game/simulation/geography/direction.py
