+++ aimmo-game/simulation/map_generator.py
@@ -6,7 +6,7 @@ from itertools import tee
 
 from six.moves import zip, range
 
-from simulation.direction import ALL_DIRECTIONS
+from simulation.geography.direction import ALL_DIRECTIONS
 from simulation.geography.location import Location
 from simulation.world_map import WorldMap
