+++ aimmo-game/tests/test_simulation/test_direction.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 
 from unittest import TestCase
 
-from simulation.direction import Direction
+from simulation.geography.direction import Direction
 
 
 class TestDirection(TestCase):
